+++ changelog.md
@@ -1,5 +1,6 @@
 0.1.4 (unreleased)
 
+* Make hooks work when repo-specific configuration file is missing
 * Improve error handling when loading custom hooks
 
 0.1.3